Refactor(feature select): separate class
SubwayApplication class -> MainView class
