+++ src/main/java/subway/view/MainView.java
@@ -1,0 +1,70 @@
+package subway.view;
+
+/**
+ * @author yhh1056
+ * @since 2020/12/11
+ */
+public class MainView {
+    private final Input input;
+    private final StationView stationView;
+    private final LineView lineView;
+    private final SectionView sectionView;
+    private final SubwayLineView subwayLineView;
+
+    public MainView(Input input) {
+        this.input = input;
+        this.stationView = new StationView(input);
+        this.lineView = new LineView(input);
+        this.sectionView = new SectionView(input);
+        this.subwayLineView = new SubwayLineView();
+    }
+
+    public void selectMainMenu() {
+        selectMenu(input.nextMainButton());
+    }
+
+    private void selectMenu(final String button) {
+        if (isEnd(button)) {
+            return;
+        }
+        selectStationMenu(button);
+        selectLineMenu(button);
+        selectSectionMenu(button);
+        selectSubwayLineMenu(button);
+
+        loopMainMenu();
+    }
+
+    private boolean isEnd(String button) {
+        return button.equals(Button.END);
+    }
+
+    private void loopMainMenu() {
+        Message.printLine();
+        selectMenu(input.nextMainButton());
+    }
+
+    private void selectStationMenu(final String button) {
+        if (button.equals(Button.ONE)) {
+            stationView.selectStationMenu(input.nextStationButton());
+        }
+    }
+
+    private void selectLineMenu(final String button) {
+        if (button.equals(Button.TWO)) {
+            lineView.selectLineMenu(input.nextLineButton());
+        }
+    }
+
+    private void selectSectionMenu(final String button) {
+        if (button.equals(Button.THREE)) {
+            sectionView.selectLineMenu(input.nextSectionButton());
+        }
+    }
+
+    private void selectSubwayLineMenu(final String button) {
+        if (button.equals(Button.FOUR)) {
+            subwayLineView.showSubwayLine();
+        }
+    }
+}
